+++ index.js
@@ -1,183 +1,12 @@
-import { createInfo, modifyInfo, optimize } from "./qbist.js"
+import { createInfo, modifyInfo } from "./qbist.js"
 import { loadStateFromParam } from "./qbistListeners.js"
+import { QbistExporter } from "./QbistExporter.js"
+import { QbistRenderer } from "./QbistRenderer.js"
 
 // UI Elements
 const loadingOverlay = document.getElementById("loadingOverlay")
 const loadingBar = document.getElementById("loadingBar")
 loadingOverlay.style.display = "none"
-
-// Renderer Management
-class QbistRenderer {
-  constructor(canvas) {
-    this.canvas = canvas
-    this.worker = null
-    this.isInitialized = false
-    this.keepAlive = false
-    console.log(`[Canvas Create] Created renderer for canvas ${canvas.id}`)
-    this._setupWorker()
-  }
-
-  _setupWorker() {
-    if (typeof Worker === "undefined") {
-      throw new Error("Web Workers are not supported in this browser")
-    }
-
-    // Cleanup existing worker if any
-    this.cleanup()
-
-    this.worker = new Worker("workerWebGL.js", { type: "module" })
-    this.worker.onerror = (err) => {
-      console.error("Worker error:", err)
-      this.cleanup()
-    }
-  }
-
-  cleanup() {
-    if (this.worker) {
-      this.worker.postMessage({ type: "cleanup" })
-      this.worker = null
-    }
-    this.isInitialized = false
-    this.keepAlive = false
-    console.log(
-      `[Canvas Delete] Cleaned up renderer for canvas ${this.canvas.id}`
-    )
-  }
-
-  async render(info, options = {}) {
-    const {
-      keepAlive = false,
-      refreshEveryFrame = false,
-      isExport = false,
-    } = options
-    this.keepAlive = keepAlive
-
-    return new Promise((resolve, reject) => {
-      if (!this.worker) {
-        this._setupWorker()
-      }
-
-      const onMessage = (e) => {
-        if (e.data.command === "rendered") {
-          if (!this.keepAlive) {
-            this.worker.removeEventListener("message", onMessage)
-          }
-          if (isExport) {
-            loadingOverlay.style.display = "none"
-          }
-          resolve(e.data)
-        } else if (e.data.command === "error") {
-          this.worker.removeEventListener("message", onMessage)
-          reject(new Error(e.data.message))
-        }
-      }
-
-      this.worker.addEventListener("message", onMessage)
-
-      try {
-        if (isExport) {
-          loadingOverlay.style.display = "flex"
-          loadingBar.style.width = "100%"
-        }
-
-        if (!this.isInitialized) {
-          // Initial setup - transfer the canvas
-          const offscreen = this.canvas.transferControlToOffscreen()
-          this.worker.postMessage(
-            {
-              type: "init",
-              canvas: offscreen,
-              info,
-              keepAlive: this.keepAlive,
-              refreshEveryFrame,
-            },
-            [offscreen]
-          )
-          this.isInitialized = true
-        } else {
-          // Just update the info for subsequent renders
-          this.worker.postMessage({
-            type: "update",
-            info,
-            keepAlive: this.keepAlive,
-            refreshEveryFrame,
-          })
-        }
-      } catch (err) {
-        this.worker.removeEventListener("message", onMessage)
-        reject(err)
-      }
-    })
-  }
-
-  update(info) {
-    if (this.worker) {
-      this.worker.postMessage({
-        type: "update",
-        info,
-        keepAlive: this.keepAlive,
-      })
-    }
-  }
-}
-
-// Export Management
-class QbistExporter {
-  constructor() {
-    this.exportCanvas = null
-    this.renderer = null
-  }
-
-  async exportImage(info, width, height) {
-    try {
-      this.cleanup()
-
-      this.exportCanvas = document.createElement("canvas")
-      this.exportCanvas.id = "exportCanvas"
-      this.exportCanvas.width = width
-      this.exportCanvas.height = height
-      document.body.appendChild(this.exportCanvas)
-      console.log(`[Canvas Create] Created export canvas ${width}x${height}`)
-
-      this.renderer = new QbistRenderer(this.exportCanvas)
-      const result = await this.renderer.render(info, { isExport: true })
-
-      if (result.kind === "bitmap") {
-        const link = document.createElement("a")
-        const tempCanvas = document.createElement("canvas")
-        tempCanvas.width = width
-        tempCanvas.height = height
-        console.log(
-          `[Canvas Create] Created temporary canvas for export ${width}x${height}`
-        )
-
-        const ctx = tempCanvas.getContext("2d")
-        ctx.drawImage(result.bitmap, 0, 0)
-
-        link.href = tempCanvas.toDataURL("image/png")
-        link.download = "qbist.png"
-        document.body.appendChild(link)
-        link.click()
-        document.body.removeChild(link)
-        console.log(`[Canvas Delete] Removed temporary canvas`)
-      }
-    } finally {
-      this.cleanup()
-    }
-  }
-
-  cleanup() {
-    if (this.renderer) {
-      this.renderer.cleanup()
-      this.renderer = null
-    }
-    if (this.exportCanvas && this.exportCanvas.parentNode) {
-      console.log(`[Canvas Delete] Removed export canvas`)
-      document.body.removeChild(this.exportCanvas)
-      this.exportCanvas = null
-    }
-  }
-}
 
 // --- Managing the 9-Panel Grid ---
 export const formulas = new Array(9)
